[c] bitwise or comparison
In C, integers are represented in bits.
A bit is dedicated to determining whether a number is negative or not.
Importantly, 1 means negative, while 0 means positive[1].

This means that when you binary or two numbers, any one of them being
negative will make the result negative.
A negative number is, by definition, less than zero, so this works out.
This should also be faster than n comparisons, making it a real
optimization.

The issue of course, is that you can't really change anything.
You can't change the comparison operator, and you definitely can't
change the 0.
Someone that takes the expression at face value is bound to litter the
code with "fancy" optimizations that act very differently from
what's expected.

[1]: I don't know if this all holds on other hardware platforms, making
this even more evil!